--- crates/radix-name-service/src/common/domain/mod.rs
@@ -1,5 +1,0 @@
-mod domain;
-mod domain_details;
-
-pub use domain::*;
-pub use domain_details::*;
