--- crates/radix-name-service/src/common/domain_record/mod.rs
@@ -1,5 +1,0 @@
-mod docket;
-mod record_details;
-
-pub use docket::*;
-pub use record_details::*;
