+++ crates/radix-name-service/src/common/mod.rs
@@ -1,7 +1,3 @@
-mod domain;
-mod domain_record;
 mod domain_to_non_fungible_id;
 
-pub use domain::*;
 pub use domain_to_non_fungible_id::*;
-pub use domain_record::*;
