+++ crates/radix-name-service/src/domain_record/record_details.rs
@@ -24,6 +24,11 @@ impl RecordDetails {
     }
 }
 
+const SCRYPTO_SBOR_RECORD_DOMAIN_ID_FIELD: &str = "domain_id";
+const SCRYPTO_SBOR_RECORD_DOCKET_CONTEXT_FIELD: &str = "context";
+const SCRYPTO_SBOR_RECORD_DOCKET_DIRECTIVE_FIELD: &str = "directive";
+const SCRYPTO_SBOR_RECORD_DOCKET_VALUE_FIELD: &str = "value";
+
 impl TryFrom<ScryptoSborValue> for RecordDetails {
     type Error = CommonError;
 
@@ -32,23 +37,25 @@ impl TryFrom<ScryptoSborValue> for RecordDetails {
             ProgrammaticScryptoSborValue::Tuple(tuple) => {
                 let domain_id = tuple
                     .fields
-                    .get_non_fungible_local_id_field("domain_id")
-                    .ok_or(CommonError::Unknown)?;
+                    .get_non_fungible_local_id_field(SCRYPTO_SBOR_RECORD_DOMAIN_ID_FIELD)
+                    .ok_or(CommonError::MissingNFTDataField { field: "Domain record domain id".to_owned() })?;
+
                 let context_str = tuple
                     .fields
-                    .get_string_field("context")
-                    .ok_or(CommonError::Unknown)?;
+                    .get_string_field(SCRYPTO_SBOR_RECORD_DOCKET_CONTEXT_FIELD)
+                    .ok_or(CommonError::MissingNFTDataField { field: "Domain record docket context".to_owned() })?;
+
                 let directive_str = tuple
                     .fields
-                    .get_enum_field("directive")
+                    .get_enum_field(SCRYPTO_SBOR_RECORD_DOCKET_DIRECTIVE_FIELD)
                     .and_then(|field| field.fields.first_string_field())
-                    .ok_or(CommonError::Unknown)?;
+                    .ok_or(CommonError::MissingNFTDataField { field: "Domain record docket directive".to_owned() })?;
 
                 let value = tuple
                     .fields
-                    .get_enum_field("value")
+                    .get_enum_field(SCRYPTO_SBOR_RECORD_DOCKET_VALUE_FIELD)
                     .and_then(|field| field.fields.first().cloned())
-                    .ok_or(CommonError::Unknown)?;
+                    .ok_or(CommonError::MissingNFTDataField { field: "Domain record value".to_owned() })?;
 
                 let context = DocketContext::from_str(&context_str)?;
                 let directive = Directive::new(directive_str);
@@ -60,7 +67,7 @@ impl TryFrom<ScryptoSborValue> for RecordDetails {
                     value,
                 })
             }
-            _ => Err(CommonError::Unknown),
+            _ => Err(CommonError::UnexpectedNFTDataFormat),
         }
     }
 }
