+++ crates/radix-name-service/src/lib.rs
@@ -1,9 +1,15 @@
 mod common;
+mod domain;
+mod domain_record;
 mod service;
+mod domain_token_receiver;
 
 pub mod prelude {
     pub use crate::common::*;
     pub use crate::service::*;
+    pub use crate::domain::*;
+    pub use crate::domain_record::*;
+    pub use crate::domain_token_receiver::*;
 
     pub(crate) use addresses::prelude::*;
     pub(crate) use bytes::prelude::*;
